+++ src/main/java/usa/badratdinova/bookexchange/repositories/BooksRepository.java
@@ -11,12 +11,12 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface BooksRepository extends JpaRepository<Book, Integer> {
+public interface BooksRepository extends JpaRepository<Book, Long> {
 
     Optional<Book> findByTitle(String title);
 
     @Query("select p from Person p left join fetch p.books b where b.id = ?1")
-    Optional<Person> findPersonByBookId(int id);
+    Optional<Person> findPersonByBookId(Long id);
 
     @Query("select b from Book b left join fetch b.person where b.title like :title%")
     List<Book> findBooksByTitle(@Param("title") String title);
